requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiosqlite==0.21.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
apscheduler==3.11.0
    # via -r requirements.in
beautifulsoup4==4.13.4
    # via -r requirements.in
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via typer
feedparser==6.0.11
    # via -r requirements.in
greenlet==3.2.3
    # via sqlalchemy
h11==0.16.0
    # via httpcore
hiredis==3.2.1
    # via redis
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   pytest-httpx
    #   respx
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.1.0
    # via pytest
loguru==0.7.3
    # via -r requirements.in
lxml==6.0.0
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
orjson==3.10.18
    # via -r requirements.in
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via pytest
pydantic==2.11.7
    # via -r requirements.in
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via
    #   pytest
    #   rich
pytest==8.4.1
    # via pytest-httpx
pytest-httpx==0.35.0
    # via -r requirements.in
python-dotenv==1.1.1
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
redis[hiredis]==6.2.0
    # via -r requirements.in
respx==0.22.0
    # via -r requirements.in
rich==14.0.0
    # via
    #   -r requirements.in
    #   typer
sgmllib3k==1.0.0
    # via feedparser
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.41
    # via -r requirements.in
typer==0.16.0
    # via -r requirements.in
typing-extensions==4.14.0
    # via
    #   aiosqlite
    #   anyio
    #   beautifulsoup4
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzlocal==5.3.1
    # via apscheduler
